i forgot everything i was doing

+++ src/chess/bit.rs
@@ -56,39 +56,12 @@ pub fn get_ternary_bitmask(piece_idx: u8, friend_mask: u8, foe_mask: u8)->u16{
 
 pub const GET_DIAG_MASK_MAGIC:u64 = 0x101010101010101;
 #[inline]
-pub fn get_pos_diag_mask(bb:u64, piece_idx: u8)->u64{
-  return ((bb & DIAG_MASK_CACHE[2*(piece_idx as usize)]) * GET_DIAG_MASK_MAGIC) >> 56;
+pub fn get_pos_diag_mask(bb:u64, piece_idx: u8)->u8{
+  return (((bb & DIAG_MASK_CACHE[2*(piece_idx as usize)]) * GET_DIAG_MASK_MAGIC) >> 56) as u8;
 }
 
 #[inline]
-pub fn get_neg_diag_mask(bb:u64, piece_idx: u8)->u64{
-  return ((bb & DIAG_MASK_CACHE[2*(piece_idx as usize) + 1]) * GET_DIAG_MASK_MAGIC) >> 56;
+pub fn get_neg_diag_mask(bb:u64, piece_idx: u8)->u8{
+  return (((bb & DIAG_MASK_CACHE[2*(piece_idx as usize) + 1]) * GET_DIAG_MASK_MAGIC) >> 56) as u8;
 }
 
-
-
-// #[inline]
-// pub fn get_bit_indices(bits: u64)-> Vec<u8> {
-//   let mut res: Vec<u8> = Vec::<u8>::with_capacity(8);
-  
-  
-
-// }
-
-
-//set bits to zero according to mask
-// w = w & !m
-
-//from: https://graphics.stanford.edu/~seander/bithacks.html#ConditionalSetOrClearBitsWithoutBranching
-//conditionally set bits without branching
-// bool f;         // conditional flag
-// unsigned int m; // the bit mask
-// unsigned int w; // the word to modify:  if (f) w |= m; else w &= ~m; 
-
-// w ^= (-f ^ w) & m;
-
-// // OR, for superscalar CPUs:
-// w = (w & ~m) | (-f & m);
-
-//find minimum of two integers without branching.
-//r = y + ((x - y) & ((x - y) >> (sizeof(int) * CHAR_BIT - 1))); // min(x, y)
